No rate display when no input nor 0 input.

+++ modules/logic/arithmetic.py
@@ -4,7 +4,7 @@ def summation(data,price,rate=0.0):
     ret = {}
     ret[u'mins'] = sum(map(lambda x: x[u'mins'], data))
     ret[u'summation'] = int(price * (ret[u'mins'] / 60.0))
-    ret[u'payment'] = ret[u'summation'] * (1 - rate)
-    ret[u'subtract'] = ret[u'summation'] * rate
+    ret[u'payment'] = int(ret[u'summation'] * (1 - rate))
+    ret[u'subtract'] = int( ret[u'summation'] * rate)
     return ret
